Updated version that disables ActiveX more thoroughly.
* patches/src680/win32-activex-disable.diff: Updated version that
disables ActiveX more thoroughly.

Add a new option to the OOo configure.in: --disable-activex. The
ActiveX Registry items in the scp2/source/*/registryitem_*.scp
files are bypassed if this option is used.

In addition, all of the scp2/source/activex/module_activex.scp and
scp2/source/activex/file_activex.scp files is bypassed. In the
previous version of this patch, the module_activex.par and
file_activex.par generated from them were dropped from
scp2/util/makefile.mk, but the end result is the same.

* bin/setup.in: On Win32, pass timestamp in --with-build-version
and --disable-activex to the OOo configure script.