As the lpsolve library gets built as a static library only (due to problems in its MSVC build, the afxres.
* patches/scsolver/scs-scp2.diff: As the lpsolve library gets
built as a static library only (due to problems in its MSVC build,
the afxres.h that lpsolve.rc tries to include isn't found), don't
try to include the nonexistant lpsolve DLL. But do fix the name of
the lpsolve DLL we would have used anyway, it would be called
lpsolve55.dll, not liblpsolve55.dll. At least that's what the
cvc6.bat tries to create, but fails. It doesn't do any harm as
such that we use a static lpsolve library, only scsolver.uno.dll
links to it.