//bugs.
* patches/src680/jvmfwk-prefer-gij-on-x86_64.diff, patches/src680/apply:
(64bitPatches) prefer gij on x86_64 with the 64-bit build; It is
useful because Sun JRE 1.5 does not work with the 64-bit build at
all. It is caused by
http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=6487640;
It is hard to debug. The wizards simply does not appear without any
message; The oobase functions ends with wery unclear message, for
example see http://forums.fedoraforum.org/showthread.php?t=135946,
n#219982