Added.
* source/ui/dialog.hxx
* source/inc/dialog.cxx (updateWidgets): added.

* source/inc/global.hxx: added SCSOLVER_DEBUG define and removed unused
global functions.

* source/ui/lpbuilder.hxx
* source/inc/lpbuilder.cxx: made constraint parsing more flexible so that
even when the LHS value is a value instead of a cell reference it would
parse correctly.

* source/numeric/lpsolve.cxx: cleaned up debug output calls.  iostream
seems broken when used with stlports.

* source/service/service.cxx: cleaned up debug output calls.

* source/ui/basedlg.cxx: extra check for NULL-ness.

* source/ui/global.cxx: replaced dysfunctional cout calls with printf
calls.

* source/ui/listener.cxx: fixed a weird behavior when the user cancels the
range selection mode.  also fixed one crasher bug when the Change or
Delete button is pressed when nothing in the constraint list box is
highlighted.

* source/ui/solvemodel.cxx: fixed constraint parsing error when constant
term is present in either side of the constraint formula.