Add novell-win32-msi-patchability.
* patches/src680/apply: Add novell-win32-msi-patchability.diff to
NovellOnlyWin32. Note that this means you will need the Win32::API
and Filesys::CygwinPaths Perl modules when building for
NovellWin32, but the speedup in instsetoo_native is definitely
worth it. Ask tml for assistance. (Those are already installed in
the zipped-up Cygwin installation at
http://go-oo.org/tml/tml-cygwin.zip .)

Merge from HEAD: Spell Fpickers consistently.