-)); Fixed the error concersing the file:// problem.
* patches/src680/sfx2-pre-and-postprocess-during-import.diff:
move a lot of copy/paste code to a static helper function; used heap
instead of stack to store argument array (suggested by michael; the
rtl_uString args[c] does not work, since the windows compiler
requires c to be a constant :-)); Fixed the error concersing the
file:// problem. So there is not need to add file:// anymore. It was
an error in the patch.