Added internal bits for enabling integer constraints on decision variables.
* source/inc/numeric/lpmodel.hxx:
* source/numeric/lpmodel.cxx:
* source/inc/option.hxx:
* source/ui/option.cxx:
* source/numeric/lpsolve.cxx: added internal bits for enabling integer
constraints on decision variables.  This is not enabled in the UI yet, but
eventually will show up in the Options dialog.

* source/tool/timer.cxx: removed #include <stdio.h>.

* source/ui/basedlg.cxx: added extra checks for NULL widget to prevent
application crash on exit when trying to unregister a NULL-pointer
listener.

* source/ui/dialog.cxx: (finally) fixed a rather serious typo here, which
was causing the whole soffice.bin process to segfault on exit every time
the Solver dialog is invoked.  I wonder why this problem did not surface
before....

* source/ui/solvemodel.cxx: added bits for integer constraints & localized
the standard exception message.