Add some Win32 specific code: Must set the environment variable MONO_CFG_DIR to the full path of the "program" folder.
* patches/src680/sfx2-pre-and-postprocess-during-import.diff: Add
some Win32 specific code: Must set the environment variable
MONO_CFG_DIR to the full path of the "program" folder. We can't
simply set MONO_CFG_DIR to ".", the current working directory when
running the converter is *not* the "program" folder, but
apparently the folder where the input file is. Get the "program"
folder by expanding $ORIGIN.

Remove code snippet in DoLoad() used to get theMacroExpander which
is not used there.