Move xalan-serializer-jar-more.
* patches/src680/apply: Move xalan-serializer-jar-more.diff to
BuildBitsLinuxOnly as it depends on cws-cmcfixes28.diff which is
in CWSBackports which is applied only on Linux. Argh, patch
interdependency is a pain when patches are conditionally applied
depending on platform (and distro). In the future we should avoid
platform-specific diffs, and instead make sure that all patches
add code that either is relevant for all platforms, or is properly
ifdeffed.

* patches/src680/apply: Add odf-converter.diff to NovellOnlyWin32.