Commit this as it actually was used by me when building 2.
* patches/src680/novell-win32-vrb-branding.diff
(desktop/util/ooverinfo.rc solenv/inc/shlinfo.rc): Commit this as
it actually was used by me when building 2.0.4-7. In retrospect it
turns out that leaving space for a patchlevel in the VRB version
is not necessary after all, MSI patching doesn't require the VRB
version to increase. And it's a pain to have to change PATCHLEVEL
in this diff file and make patch.apply.