Use OOO_PATCHLEVEL, if defined, as the fourth part in the Version column for unversioned files in the File table.
* patches/src680/novell-win32-msi-patchability.diff
(solenv/bin/modules/installer/windows/file.pm): Use
OOO_PATCHLEVEL, if defined, as the fourth part in the Version
column for unversioned files in the File table. (Otherwise use
zero, as before.)
(solenv/bin/modules/installer/windows/property.pm): Append
OOO_PATCHLEVEL, if defined, as the fourth part of the
ProductVersion in the Property table, if ProductVersion otherwise
would be just three-part (as it would).