Add win32-installer-register-moox-types.
* patches/src680/apply: Add
win32-installer-register-moox-types.diff to the Win32Only set,
too, so that it will be applied on Win32 even if the rest of the
OOXML set isn't. Thus if the person running the installer chooses
to associate OOo with MS file types, also the OOXML types will be
associated. (OOo as presently built in this branch without the
OOXML patch set won't be able to do anything with them, though,
unless the odf-converter extension is installed. But still this is
probably what we want.)