Novell Edition Windows build OOo-2.
Novell Edition Windows build OOo-2.4.0-13 at this point.

2008-03-20  Tor Lillqvist  <tml@novell.com>

* patches/src680/odf-converter-ignore-writerfilter.diff: Add code
to check if an odf-converter-* extension is deployed, and in that
case don't use the built-in OOXML filters. It is inside #ifdef WNT
as the old code in this patch that checks for program/OdfConverter
being present should be enough for the Linux case, at least when
the odf-converter executable is called OdfConverter and it is
installed through Linux package management directly into the
OOo program directory.