Novell Edition Windows build OOo-2.
Novell Edition Windows build OOo-2.4.0-15-3 at this point.

2008-04-01  Tor Lillqvist  <tml@novell.com>

* patches/src680/win32-revert-solenv-2.3.0.diff: Revert the Perl
code that constructs the MSI database to the 2.3.0 version, and
some further hacks so that it actually works together with other
stuff that has changed here in 2.4.0. This means we can have a
shitload of languages in our multi-lingual installer again, yay!

This diff hasn't been tested as such in a from-scratch build, but
was constructed after making a backup copy of solenv and manually
reverting solenv/bin/make_installer.pl and solenv/bin/modules from
2.3.0, and then doing further manual edits. So if/when a fresh
from scratch build is done, will have check carefully that this
patch actually works.

* patches/src680/apply: Use it.