Just drop the ALL_LANG(Tooltip, STR_FI_TOOLTIP_FOO) entries from the scp files, so that no tooltips are created in the first place, instead of trying to make the tooltips less lame and less misleading.
* patches/dev300/win32-tooltips.diff: Just drop the
ALL_LANG(Tooltip, STR_FI_TOOLTIP_FOO) entries from the scp files,
so that no tooltips are created in the first place, instead of
trying to make the tooltips less lame and less misleading. They
are localised anyway, and I don't want to attempt to edit them in
umpteen languages in the localize.sdf file.