2008-11-11  Tor Lillqvist  <tml@novell.com>
* patches/dev300/novell-win32-msi-patchability.diff
(instsetoo_native/inc_openoffice/windows/msi_templates/InstallE.idt):
Change the condition of the SetOfficeInstallLocation,
SetBasisInstallLocation and SetUreInstallLocation actions in the
InstallExecuteSequence table to "Not Installed Or PATCH" instead
of just "Not Installed". This precents the Registry entries that
refer to [OFFICEINSTALLLOCATION] etc from getting screwed up when
patching. Just a stopgap measure, not to be committed to
trunk. Upstream will fix this in a slightly different way in 3.1 I
think.