Move jpegc.
* patches/dev300/apply: Move jpegc.c.diff from Fixes to
LinuxOnly. As its comment says, it is intended only for use with
an *unpatched* libjpeg. "unpatched" here means that the upstream
jpeg-6b.patch has not been applied to it. Using jpegc.c.diff on
Windows corrupts colours in CMYK or YCCK JPEGs, bnc#444106. BTW, I
wonder what YCCK is? Presumably this patch should not be used on
MacOS and Solaris either, as I think they use the included patched
libjpeg, too.