New file.
* patches/dev300/mail-document-subject.diff
(sfx2/source/dialog/mailmodel.cxx): New file. Tentative fix for
bnc#459176. If no subject has been set, use the basename of the
first attached document as the subject, and in case there are
several attached documents, append a command and ellipsis. I don't
know if it is actually possible to send several documents as mail
from OOo at the same time, though.

Not yet added to the apply file.