Fixed a regression where it was impossible to create new percent-stacked charts.
* patches/dev300/chart-axis-multi-chart-types-xmloff.diff:
* patches/dev300/chart-axis-multi-chart-types-chart2.diff:
* patches/dev300/apply: fixed a regression where it was impossible to
create new percent-stacked charts.  Note that I still need to support
saving of charts with mixed stack modes, because the current ODF format
only allows a single stack mode per diagram.  Because of this, when you
import an xls chart with mixed stack modes and save it to ODF, you'll
lose that information. (n#472264)  These patches are disabled in the
default build.