Added support for allowing the users to specify an alternative language to use when importing an HTML file into Calc.
* patches/dev300/calc-html-import-custom-lang-filter.diff:
* patches/dev300/calc-html-import-custom-lang-sc.diff:
* src/sdf/sc-langchooser-fr.sdf:
* src/sdf/sc-langchooser-ja.sdf:
* patches/dev300/apply: added support for allowing the users to specify
an alternative language to use when importing an HTML file into Calc.
This is useful when importing an HTML table containing numbers with dots
as the decimal separators (e.g. 12.56) into a document that uses commas
as the decimal separators (e.g. 12,56), and so on. (n#484272)

NOTE: this is not enabled in the build unless CalcExperimental is 
included.