Move win32-multi-lang-installer.diff patch to Novell related sections
Created MultilanguageWin32Only section in apply file and put
win32-multi-lang-installer.diff patch into, because this patch
causes not running installations when the installed OOo's language is
not match with the language of Windows of at least the regional settings of
Windows. I put this newly created section into Novell and Go-OO versions.

* patches/dev300/apply: