Docx: reused hdr/ftr now imported
n519201: in Docx, when no header/footer reference is contained in a
sectPr, the ones of the previous section is used. This is now
implemented.

In order to copy the previous header/footer text into the new one, the
SwXText object had to support the XTextCopy interface, but it doesn't
tells it to the world: this is not a published API.

* patches/dev300/docx-import-fixes.diff: