Don't make .msp file name too long.
There is no need to put the language string in the name of the .msp
file being built. The folder name already contains the language string
(or, as it happens, a hash code of it in the cases where the number of
languages is large, as it is for us (Novell), to avoid too long
pathnames). The code used the language string as such as part of the
msp file name, which made it much too long for us.