Use the white artwork for Novell Edition here, too.
Copy the openintro_nld.svg file from the 3.1 branch and change the
version to 3.1.1. Copy also openintro_nld.bmp and openabout_nld.bmp
from the 3.1 branch. As the SVG uses the Swiss721 typeface which I
don't have I could not re-render the openintro_nld.bmp, but luckily in
this case it was easy to just edit it in GIMP and add a copy of the
".1" to the version number...