Clean up DejaVu font mess
Upstream now includes DejaVu fonts version 2.30. Clean up the
ooo-build infrastructure related to using different DejaVu
versions.

For now we presumably no longer need to specify a different DejaVu
version as the upstream one is quite recent.

Require the --with-dejavu-fonts option, if used, to specify a version
number larger than the upstream one.

Don't couple the use of Liberation fonts to the --with-dejavu-fonts
option.

Rename the DejaVuFonts patch set to LiberationFontsOnWindows, which
says what it actually does.