Fix a problem in WebDAV code
Small change that seems to fix bnc#553819. Use the local xResAccess
variable instead of the member m_xResAccess in one more place inside
Content::open(), like in other places. This avoids a crash caused by
heap corruption, valgrind reports access to freed data. Unfortunately
I was not able to wrap my head around the code enough to understand
the exact mecahism which causes that, though. But as long as this
fixes the crash, I am satisfied.