Solve the pdfimport problem in a better way
Just always copy the stream into a temporary file and load it from
there. Then we don't nee to bypass the check for an encrypted
document, that should work now. Remove the temp file after it has been
loaded to avoid leaving potentially discriminating evidence around.

Thanks to Fridrich for the idea.