Don't treat x86_64 and i386 differently
* patches/dev300/cairo-pixman-makefile-mk.diff: build shared pixman
for all linux configurations. It is possible to patch out the packaging
of cairo and pixman and rely at the end on the fact that modern Linux
distributions have a compatible cairo version. But since almost no Linux
distribution apart GoOoLinux uses internal cairo, it will be a bit moot
question for them.