sw: preserve character style on 'reset format'
It changes the behavior of "Default formatting" option to
NOT remove the attribute for character style. It will only
remove directly applied format.

It will also NOT remove hyperlinks anymore. People should
use the Remove Hyperlink command.

For those that would need to set all the marked text to
"Default" style, they can still do it through F11 > Default.

See discussion in i#47893 and i#85464 and the corresponding
duplicates.

* patches/dev300/apply:
* patches/dev300/sw-preserve-char-style-on-reset-format.diff:
  sw/source/ui/shells/textsh1.cxx:
    in func void SwTextShell::Execute(SfxRequest &rReq)
      in case FN_FORMAT_RESET

Signed-off-by: Octavio Alvarez <alvarezp@alvarezp.ods.org>