Allow PDF export on filtered range selection.
When exporting a selection to PDF, we should not split the selected
range by the filtered or hidden ranges but use the whole selected
range as a single range.  Otherwise exporting such ranges would
generate an empty PDF output.

Also, when the user is trying to export multiple selections,
throw an error message, instead of failing silently with an empty
PDF output. (n#585028)

* patches/dev300/apply:
* patches/dev300/calc-pdf-export-allow-filtered-range-filter.diff:
* patches/dev300/calc-pdf-export-allow-filtered-range-sc.diff: