Re-establish DDE linkage upon loading source documents.
When opening a source document that is the server of a DDE
linkage of another document, automatically re-establish
linkage to allow automatic updates.  This patch also fixes
the source document opeing read-only, and also a bug in
reference parsing when the source document was a Calc
document.

This corresponds with n#618846 and n#618864.  Disabled in the
default build.  To use, enable [ CalcExperimental ] section.

* patches/dev300/apply:
* patches/dev300/dde-reconnect-on-load-sc.diff:
* patches/dev300/dde-reconnect-on-load-sfx2.diff:
* patches/dev300/dde-reconnect-on-load-svx.diff:
* patches/dev300/dde-reconnect-on-load-sw.diff: