Update the build directory when the repos changed.
When there were patches moved to the repos, we need to make sure that we are
still able to apply patches; so we need to unapply all the patches, rsync
changes from rawbuild/ to build/oooABC-mX, and apply the patches again.

Unfortunately, it can take a while, but works well.