Support for pre-action flags in bin/g
the first pre-action flag supported is -f, that tell
bin/g to apply the git command (which ever it is) to every git repository
regardless of the completion of that git command for each repository.
(the default, except for pull, status and grep is to stop after the
first 'error'

This was motivated by the need to do bin/g -f stash pop.