git-hooks: Rewrote most of the pre-commit hook in perl.
This was to incorporate the changes from fix-trailing-whitespace.pl,
otherwise we had problems with locating that script :-(  Separate
fix-trailing-whitespace.pl is not necessary any more, removed that.

Some pieces still stay as a bigger shell pieces, I did not care enough
to change those - feel free to improve it, please.