No extra ACL check when loading documents.
Performing extra ACL check when loading document caused the
Document in Use dialog NOT to appear when the document is
accessed by a 2nd user and the document is located in a
samba share.  This extra ACL check was originally placed as
a temporary fix for i#102464, but it doesn't appear to be
needed any more. (n#653747)

* patches/dev300/apply:
* patches/dev300/file-load-no-extra-acl-check.diff: