Work on WebDAV locking
The code already contains the ucb part (from OOo), and it mostly seems
to work fine.

But, the sfx2 part is missing. OOo is working on that in the CWS
mav58. Introduce the code from that in cws-mav58.diff, an extract of
that CWS, modified to apply.

The old webdav-locking.diff can thus finally be retired, good
riddance.

Re-insert the sfx2 pre and postprocessing code for saving into the
code as now modified by cws-mav58.diff.

The Apache server handling the Novell Innerweb Vibe OnPrem service
(formerly known as "Teaming") decreases the WebDAV timeout granted by
one second from what was requested each time a lock is granted or
refreshed.

As we start with a requested timeout of 180 seconds, and refresh the
lock always 30 seconds before it is about to expire, this means that
after about 150*180 seconds (7.5 hours) the timeout would reach zero
and locking would break. Which is likely to happen, as people might
well keep a document open in LibreOffice for the whole day.

Avoid this by not reducing the timeout requested next time even if the
server granted a slightly shorter timeout than what we asked for.

There are still problems. The most obvious one is that after saving a
document into WebDAV once, it no longer is locked... Somehow the code
seems to forget the authentication to use, but only for lock requests,
huh? Possibly it needlessly tears down the WebDAV session completely
after saving the document. Further debugging needed. Should
investigate why it can't just keep the same session live for the same
document.

Build neon with debugging printout capability, and ask for neon
debugging printout in the webdav UCP if built with debug=t.