Make the now non-empty WebDAV set NovellOnly until approved
I want to be able to commit work on this in pieces at my own pace, so
don't "force" it upon other distros yet. Once I am reasonably certain
that it is good enough, I will ask for approval to put it in Common.

I can't really do this in master, as I want this for the Novell 3.3.1
or 3.3.2 build.

The patches in WebDAVUpstream is better classified as
UnUsedButNotYetRemovedFromGit. Actually I guess I should just remove
them.