Kill i92372.diff
It has been commented out over a year. It was a very ad-hoc fix for a
crash when embeddeing OOo as a message editor in GroupWise. Let's hope
GroupWise doesn't even try to use LibreOffice as an embedded editor
any more, or that the crash is gone thanks to other changes that have
happened since I cooked up this patch.